fix: share preset finish tool hook helper
